+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { UserAccount } from './components/UserAccount'
+import { UserAccount } from './components/UserAccount/UserAccount'
 
 const App: React.FC = () => {
     return <UserAccount />
